re #186 testing. rename '1lm' feature type to 'single landmark'

+++ TrainingOverlayReceiver.m
@@ -18,9 +18,9 @@ classdef TrainingOverlayReceiver < handle
         end        
       else
         trow = obj.tblMFT(eid,:);
-        for i=1:numel(obj.axs)
+        for i=1:numel(obj.axs)          
           tstr = sprintf('%s \\color{darkgreen}Selection: (mov %d,frm %d,tgt %d)',...
-            obj.tstrBases{i},trow.mov,trow.frm,trow.iTgt);
+            obj.tstrBases{i},int32(trow.mov),trow.frm,trow.iTgt);
           title(obj.axs(i),tstr,'fontweight','bold');
         end
       end
